Added ObjectHandler.cs
Only use this if you know what you are doing! This is for those of us
who want to speed improve their assemblies and can't wait for Jodus to
apply his changes to the actual LeagueSharp.dll.
Once he improves the ObjectManager itself this will be removed, this is
why it's marked as obsolete!